Refactoring: moved functions to cli-utils
In order to be able to test them.
